Fix a warning

+++ dev/AppLifecycle/ExtensionContract.h
@@ -55,7 +55,7 @@ namespace winrt::Microsoft::Windows::AppLifecycle::implementation
         // QueryParsed() function would return empty when it's a file contract with unicode characters in file path
         // Thus following additional check for file contract is needed
         auto fileContractUri = GenerateEncodedLaunchUri(L"App", c_fileContractId);
-        if (CompareStringOrdinal(uri.AbsoluteUri().c_str(), fileContractUri.length(), fileContractUri.c_str(), -1, TRUE) == CSTR_EQUAL)
+        if (CompareStringOrdinal(uri.AbsoluteUri().c_str(), static_cast<int>(fileContractUri.length()), fileContractUri.c_str(), -1, TRUE) == CSTR_EQUAL)
         {
             return { ExtendedActivationKind::File, FileActivatedEventArgs::Deserialize(uri) };
         }
